Unify & shipinit MiscBehaviour
