+++ src/port/Enhancements/Events/PortEnhancements.cpp
@@ -112,7 +112,6 @@ void PortEnhancements_Register() {
     REGISTER_EVENT(OnActorUpdate);
 
     // Register rando events
-    REGISTER_EVENT(InitRandoEvents);
     REGISTER_EVENT(OnLoadFileSelect);
     REGISTER_EVENT(OnSaveLoad);
     REGISTER_EVENT(OnWarpDispatch);
